requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.2.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
Pillow==10.4.0  # Updated to latest version
requests==2.31.0
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
stripe==7.8.0
django-environ==0.11.2